backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
PyMySQL
django-cors-headers==4.4.0

Pillow>=11.0
django-filter==24.2
python-decouple==3.8
